backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
drf-spectacular==0.27.1
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.2.0
reportlab==4.0.9
pytest==7.4.4
pytest-django==4.7.0
black==24.1.1
flake8==7.0.0
isort==5.13.2
faker==22.6.0

